requirements.txt
mcp[cli]
watchdog
requests
openpyxl
python-dotenv
pyyaml
Pillow
google-generativeai